+++ src/iexec-workerpool.js
@@ -80,12 +80,12 @@ deploy
       }
       await keystore.load();
       spinner.start(info.deploying(objName));
-      const address = await hub.deployWorkerpool(
+      const { address, txHash } = await hub.deployWorkerpool(
         chain.contracts,
         iexecConf[objName],
       );
       spinner.succeed(`Deployed new ${objName} at address ${address}`, {
-        raw: { address },
+        raw: { address, txHash },
       });
       await saveDeployedObj(objName, chain.id, address);
     } catch (error) {